backend/requirements.txt
fastapi
uvicorn
yt-dlp
google-genai
supabase
python-dotenv
ffmpeg-python
pydantic
